requirements.txt
aniso8601==4.0.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
nltk==3.4
python-dotenv==0.10.1
pytz==2018.7
requests==2.20.0
singledispatch==3.4.0.3
six==1.12.0
urllib3==1.24
Werkzeug==0.14.1
